+++ lua/lazyvim/util/plugin.lua
@@ -17,6 +17,7 @@ M.deprecated_extras = {
   ["lazyvim.plugins.extras.ui.dashboard"] = "`dashboard.nvim` is now the default **LazyVim** starter.",
   ["lazyvim.plugins.extras.coding.native_snippets"] = "Native snippets are now the default for **Neovim >= 0.10**",
   ["lazyvim.plugins.extras.ui.treesitter-rewrite"] = "Disabled `treesitter-rewrite` extra for now. Not ready yet.",
+  ["lazyvim.plugins.extras.ui.treesitter-main"] = "the `nvim-treesitter` main branch is now used by default",
   ["lazyvim.plugins.extras.coding.mini-ai"] = "`mini.ai` is now a core LazyVim plugin (again)",
   ["lazyvim.plugins.extras.lazyrc"] = "local spec files are now a lazy.nvim feature",
   ["lazyvim.plugins.extras.editor.trouble-v3"] = "Trouble v3 has been merged in main",