+++ lua/lazyvim/util/ui.lua
@@ -1,22 +1,22 @@
 ---@class lazyvim.util.ui
 local M = {}
 
--- optimized treesitter foldexpr
+M.installed = {} ---@type string[]
+
+---@param ft string
+function M.have(ft)
+  local lang = vim.treesitter.language.get_lang(ft)
+  return vim.tbl_contains(M.installed, lang)
+end
+
 function M.foldexpr()
   local buf = vim.api.nvim_get_current_buf()
-  if vim.b[buf].ts_folds == nil then
-    -- as long as we don't have a filetype, don't bother
-    -- checking if treesitter is available (it won't)
-    if vim.bo[buf].filetype == "" then
-      return "0"
-    end
-    if vim.bo[buf].filetype:find("dashboard") then
-      vim.b[buf].ts_folds = false
-    else
-      vim.b[buf].ts_folds = pcall(vim.treesitter.get_parser, buf)
-    end
-  end
-  return vim.b[buf].ts_folds and vim.treesitter.foldexpr() or "0"
+  return M.have(vim.b[buf].filetype) and vim.treesitter.foldexpr() or "0"
+end
+
+function M.indentexpr()
+  local buf = vim.api.nvim_get_current_buf()
+  return M.have(vim.b[buf].filetype) and require("nvim-treesitter").indentexpr() or -1
 end
 
 return M